Move array related utils to array_util
