+++ src/bots/discord.ts
@@ -6,7 +6,8 @@ import Command from '../commands/command';
 import ConfigManager from '../managers/config_manager';
 import Notification from '../notifications/notification';
 import MDRegex from '../util/regex';
-import { StrUtil, mapAsync, optMapAsync } from '../util/util';
+import { mapAsync, optMapAsync } from '../util/array_util';
+import { StrUtil } from '../util/util';
 import Message from '../message';
 import Permissions from '../permissions';
 import ProjectManager from '../managers/project_manager';
