+++ src/bots/telegram.ts
@@ -7,7 +7,8 @@ import Command from '../commands/command';
 import ConfigManager from '../managers/config_manager';
 import Notification from '../notifications/notification';
 import MDRegex from '../util/regex';
-import { StrUtil, mapAsync } from '../util/util';
+import { mapAsync } from '../util/array_util';
+import { StrUtil } from '../util/util';
 import Message from '../message';
 import Permissions from '../permissions';
 import Game from '../game';
