+++ src/commands/commands.ts
@@ -5,7 +5,8 @@ import EscapeRegex from 'escape-string-regexp';
 import { UserRole } from '../user';
 import getBots from '../bots/bots';
 import Game from '../game';
-import { mapAsync, matchGroups, naturalJoin } from '../util/util';
+import { mapAsync, naturalJoin } from '../util/array_util';
+import { matchGroups } from '../util/util';
 import ProjectManager from '../managers/project_manager';
 import CommandGroup from './command_group';
 import SimpleAction from './simple_action';
