+++ tests/util.spec.ts
@@ -1,62 +1,6 @@
-import { mapAsync, filterAsync, naturalJoin, ObjUtil, StrUtil } from '../src/util/util';
+import { ObjUtil, StrUtil } from '../src/util/util';
 
 describe('Util', () => {
-  describe('map async', () => {
-    test('works', async () => {
-      const testArray = [1, 2, 3];
-      const asyncFunction = async (el: number) => el + 1;
-      const rr = await mapAsync(testArray, asyncFunction);
-      expect(rr).toEqual([2, 3, 4]);
-    });
-  });
-
-  describe('filter async', () => {
-    test('works', async () => {
-      const testArray = [1, 2, 3];
-      const asyncFunction = async (el: number) => el > 2;
-      const rr = await filterAsync(testArray, asyncFunction);
-      expect(rr).toEqual([3]);
-    });
-  });
-
-  describe('natural join', () => {
-    test('undefined', () => {
-      const testArray = undefined;
-      const result = naturalJoin(testArray);
-      expect(result).toEqual('');
-    });
-
-    test('empty', () => {
-      const testArray: string[] = [];
-      const result = naturalJoin(testArray);
-      expect(result).toEqual('');
-    });
-
-    test('1 item', () => {
-      const testArray: string[] = ['one'];
-      const result = naturalJoin(testArray);
-      expect(result).toEqual('one');
-    });
-
-    test('2 items', () => {
-      const testArray: string[] = ['one', 'two'];
-      const result = naturalJoin(testArray);
-      expect(result).toEqual('one and two');
-    });
-
-    test('3 items', () => {
-      const testArray: string[] = ['one', 'two', 'three'];
-      const result = naturalJoin(testArray);
-      expect(result).toEqual('one, two and three');
-    });
-
-    test('3 items with separator', () => {
-      const testArray: string[] = ['one', 'two', 'three'];
-      const result = naturalJoin(testArray, '|');
-      expect(result).toEqual('one|two and three');
-    });
-  });
-
   describe('Object util', () => {
     describe('keys', () => {
       test('array', () => {
